Saft data parser
Saft is a data representation format similar to JSON with slight tweaks
making it more suitable to represent some types of data.

It's the evolution of https://github.com/johan-bolmsjo/pot with the
addition of raw strings, comments and some questionable string escaping
possibilities removed. Raw strings are suitable to represent regexps in
configuration files to avoid double quoting.

This parser is planed to be used to parse the configuration file for a
command line tool yet to be completed. It's possible that some auxiliary
parsing support will be developed at that time as the API in this
package is rather minimal. In that case that will probably become a
separate package.

See README.md for format syntax.